Refactored HourRateController
