+++ packages/api/AlvTimeWebApi/ServiceRegistrator.cs
@@ -39,7 +39,7 @@ namespace AlvTimeWebApi
             services.AddScoped<TaskCreator>();
             services.AddScoped<ITimeRegistrationStorage, TimeRegistrationStorage>();
             services.AddScoped<IHourRateStorage, HourRateStorage>();
-            services.AddScoped<HourRateCreator>();
+            services.AddScoped<HourRateService>();
             services.AddScoped<IProjectStorage, ProjectStorage>();
             services.AddScoped<ProjectCreator>();
             services.AddScoped<ICustomerStorage, CustomerStorage>();
